#157: Set OpenSHR data to be persistence

+++ run.sh
@@ -25,28 +25,31 @@ else
 	echo "Copying module dependencies and reference application modules..."
 	mkdir -pv $OPENMRS_MODULES
 	cp /root/temp/modules/*.omod $OPENMRS_MODULES
-	echo "Modules copied."
+	echo "Modules copied."	
 
-	echo "### Init database for OpenMRS ###"
-	mysql -u root -pshr -e "DROP DATABASE openmrs;"
+	DB=`mysql -u root -pshr --skip-column-names -e "SHOW DATABASES LIKE '${OPENMRS_DATABASE}'"`
+	if [ "$DB" != "${OPENMRS_DATABASE}" ]; then
+		echo "# Init database for OpenMRS #"
 
-	echo "Delete exist database OpenMRS"
-	mysql -u root -pshr -e "CREATE DATABASE openmrs;"
+		mysql -u root -pshr -e "CREATE DATABASE ${OPENMRS_DATABASE};"
 
-	echo "Create database openmrs from file < openmrs.sql"
-	mysql -u root -pshr openmrs < openmrs.sql
+		mysql -u root -pshr -e "GRANT ALL ON openmrs.* to '${MYSQL_USER}'@'%' identified by '${MYSQL_PASSWORD}';"
 
-	echo "Update database openmrs with SHR configuration from file < openshr-configuration.sql"
-	mysql -u root -pshr openmrs < openshr-configuration.sql
+		echo "Create database openmrs from file < openmrs.sql"
+		mysql -u root -pshr openmrs < openmrs.sql
 
+		echo "Update database openmrs with SHR configuration from file < openshr-configuration.sql"
+		mysql -u root -pshr openmrs < openshr-configuration.sql
 
-	echo "Finish init database"
-	rm openmrs.sql	
-	rm openshr-configuration.sql
-	sleep 5
+		echo "Finish init database"
+		rm openmrs.sql	
+		rm openshr-configuration.sql
+	else
+		echo "# Database already exists. Skipping database creation and import. #"
+	fi
 
 fi
 
 
-echo "### Start TOMCAT ###"
+echo "# Start TOMCAT #"
 catalina.sh run
